backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
email-validator
google-generativeai
